Update STHbench-Dev012.11.sh
Added removeBenches() and consolidated the removals there. Called remove benches just after the echo "Uninstall benches" near the end of the file.

Some of the modules were not going back to the bench directory before processing the rm -rf commands. As a result, we were not removing them consistently.

Also, since they are removed out of the main modules that may make a few easier to loop.

We can not use this change if we do not want, but figured I would add.