Update linux-bench.sh
The strtoll library function is returning zero content length which makes Curl_readwrite in lib/transfer.c quit. It works for transfer -encoding type chunked as there it ignores content length and keep reading so this explains why it works for few sites and fails for rest. Implemented custom solution.